requirements.txt
simpleaudio
line-bot-sdk
python-dotenv
pygame
requests
openai